backend/requirements.txt
fastapi
uvicorn
google-cloud-firestore
google-cloud-aiplatform
vertexai
python-dotenv
pydantic
httpx
